+++ examples/simple_react/src/components/signOut/style.ts
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const style = {
-  user: {
-    lineHeight: 2.1,
-  } as React.CSSProperties,
-};
-
-export default style;