+++ packages/backend-auth/src/factory.ts
@@ -14,22 +14,26 @@ import {
   ResourceProvider,
 } from '@aws-amplify/plugin-types';
 import * as path from 'path';
-import { AuthLoginWithFactoryProps } from './types.js';
+import { AuthLoginWithFactoryProps, Expand } from './types.js';
 import { translateToAuthConstructLoginWith } from './translate_auth_props.js';
 
-export type TriggerConfig = {
-  triggers?: Partial<
-    Record<TriggerEvent, ConstructFactory<ResourceProvider<FunctionResources>>>
-  >;
-};
-
-export type AmplifyAuthFactoryProps = Omit<
-  AuthProps,
-  'outputStorageStrategy' | 'loginWith'
-> &
-  TriggerConfig & {
-    loginWith: AuthLoginWithFactoryProps;
-  };
+export type AmplifyAuthProps = Expand<
+  Omit<AuthProps, 'outputStorageStrategy' | 'loginWith'> & {
+    /**
+     * Specify how you would like users to log in. You can choose from email, phone, and even external providers such as LoginWithAmazon.
+     */
+    loginWith: Expand<AuthLoginWithFactoryProps>;
+    /**
+     * Configure custom auth triggers
+     */
+    triggers?: Partial<
+      Record<
+        TriggerEvent,
+        ConstructFactory<ResourceProvider<FunctionResources>>
+      >
+    >;
+  }
+>;
 
 /**
  * Singleton factory for AmplifyAuth that can be used in Amplify project files
@@ -44,7 +48,7 @@ class AmplifyAuthFactory
    * Set the properties that will be used to initialize AmplifyAuth
    */
   constructor(
-    private readonly props: AmplifyAuthFactoryProps,
+    private readonly props: AmplifyAuthProps,
     private readonly importStack = new Error().stack
   ) {}
 
@@ -72,7 +76,7 @@ class AmplifyAuthGenerator implements ConstructContainerEntryGenerator {
   private readonly defaultName = 'amplifyAuth';
 
   constructor(
-    private readonly props: AmplifyAuthFactoryProps,
+    private readonly props: AmplifyAuthProps,
     private readonly getInstanceProps: ConstructFactoryGetInstanceProps
   ) {}
 
@@ -103,9 +107,10 @@ class AmplifyAuthGenerator implements ConstructContainerEntryGenerator {
 }
 
 /**
- * Creates a factory that implements ConstructFactory<AmplifyAuth & ResourceProvider<AuthResources>>
+ * Provide the settings that will be used for authentication.
  */
 export const defineAuth = (
-  props: AmplifyAuthFactoryProps
+  props: AmplifyAuthProps
 ): ConstructFactory<AmplifyAuth & ResourceProvider<AuthResources>> =>
+  // Creates a factory that implements ConstructFactory<AmplifyAuth & ResourceProvider<AuthResources>>
   new AmplifyAuthFactory(props, new Error().stack);