+++ packages/backend-data/src/validate_authorization_modes.ts
@@ -1,4 +1,4 @@
-import { AuthorizationModes as CDKAuthorizationModes } from '@aws-amplify/graphql-api-construct';
+import { AuthorizationModes as CDKAuthorizationModes } from '@aws-amplify/data-construct';
 import { AuthorizationModes } from './types.js';
 
 type AuthorizationModeValidator = (