+++ packages/cli/src/commands/generate/config/generate_config_command.test.ts
@@ -5,8 +5,10 @@ import { ClientConfigFormat } from '@aws-amplify/client-config';
 import yargs, { CommandModule } from 'yargs';
 import { TestCommandRunner } from '../../../test-utils/command_runner.js';
 import assert from 'node:assert';
-import { BackendIdentifierResolver } from '../../../backend-identifier/backend_identifier_resolver.js';
+import { AppBackendIdentifierResolver } from '../../../backend-identifier/backend_identifier_resolver.js';
 import { ClientConfigGeneratorAdapter } from '../../../client-config/client_config_generator_adapter.js';
+import { BackendIdentifierResolverWithFallback } from '../../../backend-identifier/backend_identifier_with_sandbox_fallback.js';
+import { SandboxBackendIdResolver } from '../../sandbox/sandbox_id_resolver.js';
 
 void describe('generate config command', () => {
   const clientConfigGeneratorAdapter = new ClientConfigGeneratorAdapter(
@@ -19,9 +21,22 @@ void describe('generate config command', () => {
     () => Promise.resolve()
   );
 
-  const backendIdResolver = new BackendIdentifierResolver({
+  const namespaceResolver = {
     resolve: () => Promise.resolve('testAppName'),
-  });
+  };
+
+  const defaultResolver = new AppBackendIdentifierResolver(namespaceResolver);
+
+  const sandboxIdResolver = new SandboxBackendIdResolver(namespaceResolver);
+  const fakeSandboxId = 'my-fake-app-my-fake-username';
+  const mockedSandboxIdResolver = mock.method(sandboxIdResolver, 'resolve');
+  mockedSandboxIdResolver.mock.mockImplementation(() => fakeSandboxId);
+
+  const backendIdResolver = new BackendIdentifierResolverWithFallback(
+    defaultResolver,
+    sandboxIdResolver
+  );
+
   const generateConfigCommand = new GenerateConfigCommand(
     clientConfigGeneratorAdapter,
     backendIdResolver
@@ -33,6 +48,16 @@ void describe('generate config command', () => {
 
   beforeEach(() => {
     generateClientConfigMock.mock.resetCalls();
+  });
+
+  void it('uses the sandbox id by default if stack or branch are not provided', async () => {
+    const handlerSpy = mock.method(
+      clientConfigGeneratorAdapter,
+      'generateClientConfigToFile'
+    );
+    await commandRunner.runCommand('config');
+
+    assert.equal(handlerSpy.mock.calls[0].arguments[0], fakeSandboxId);
   });
 
   void it('generates and writes config for stack', async () => {