+++ packages/cli/src/commands/pipeline-deploy/pipeline_deploy_command.ts
@@ -67,6 +67,7 @@ export class PipelineDeployCommand
 
   builder = (yargs: Argv): Argv<PipelineDeployCommandOptions> => {
     return yargs
+      .version(false)
       .option('branch', {
         describe: 'Name of the git branch being deployed',
         demandOption: true,