+++ packages/integration-tests/test-projects/standalone-data-auth-modes/amplify/data/resource.ts
@@ -15,7 +15,7 @@ export const data = defineData({
     }
   `,
   authorizationModes: {
-    defaultAuthorizationMode: 'AWS_LAMBDA',
+    defaultAuthorizationMode: 'lambda',
     lambdaAuthorizationMode: {
       function: Func.fromDir({
         name: 'ApiAuth',