+++ packages/storage-construct/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aws-amplify/storage-construct-alpha
 
+## 0.2.3
+
+### Patch Changes
+
+- 5ed51cbd5: Upgrade aws-cdk to 2.110.1
+- Updated dependencies [5ed51cbd5]
+  - @aws-amplify/backend-output-storage@0.2.6
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [07b0dfc9f]
+  - @aws-amplify/backend-output-schemas@0.4.0
+  - @aws-amplify/backend-output-storage@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes